requirements.txt
ultralytics>=8.0.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
pyttsx3>=2.90
PyYAML>=6.0
lap>=0.4.0
requests>=2.28.0
